# swagger/room.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "MeetUp Project"
  termsOfService: "http://localhost:8080"
  contact:
    email: "juarez.lucasdaniel@gmail.com"
host: "localhost:8080"
basePath: "/meetUp"
tags:
- name: "Room"
  description: ""
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "https"
- "http"
paths:
  /room:
    post:
      tags:
      - "Room"
      summary: "Crear nueva sala."
      description: ""
      operationId: "addRoom"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Room"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Room"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/FormatError"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/FormatError"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/FormatError"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/FormatError"
        "405":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/FormatError"
        "408":
          description: "Request Tiemout"
          schema:
            $ref: "#/definitions/FormatError"
        "418":
          description: "Controlled Error"
          schema:
            $ref: "#/definitions/FormatError"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/FormatError"
      security:
      - room_auth:
        - "write:room"
        - "read:room"
  /room/{roomId}:
    put:
      tags:
      - "Room"
      summary: "Modificar una sala."
      description: ""
      operationId: "updateRoom"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "roomId"
        in: "path"
        description: "Identificador de la sala"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/UpdateRoom"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Room"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/FormatError"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/FormatError"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/FormatError"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/FormatError"
        "405":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/FormatError"
        "408":
          description: "Request Tiemout"
          schema:
            $ref: "#/definitions/FormatError"
        "418":
          description: "Controlled Error"
          schema:
            $ref: "#/definitions/FormatError"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/FormatError"
      security:
      - room_auth:
        - "write:room"
        - "read:room"
    get:
      tags:
      - "Room"
      summary: "Consultar una sala."
      description: ""
      operationId: "getRoom"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "roomId"
        in: "path"
        description: "Identificador de la sala"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Room"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/FormatError"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/FormatError"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/FormatError"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/FormatError"
        "405":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/FormatError"
        "408":
          description: "Request Tiemout"
          schema:
            $ref: "#/definitions/FormatError"
        "418":
          description: "Controlled Error"
          schema:
            $ref: "#/definitions/FormatError"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/FormatError"
      security:
      - room_auth:
        - "write:room"
        - "read:room"
    delete:
      tags:
      - "Room"
      summary: "Eliminar una sala."
      description: ""
      operationId: "deleteRoom"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "roomId"
        in: "path"
        description: "Identificador de la sala"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful Operation"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/FormatError"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/FormatError"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/FormatError"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/FormatError"
        "405":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/FormatError"
        "408":
          description: "Request Tiemout"
          schema:
            $ref: "#/definitions/FormatError"
        "418":
          description: "Controlled Error"
          schema:
            $ref: "#/definitions/FormatError"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/FormatError"
      security:
      - room_auth:
        - "write:room"
        - "read:room"
securityDefinitions:
  room_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:8080/oauth/token"
    flow: "implicit"
    scopes:
      write:room: "modify room in your account"
      read:room: "read your room"
definitions:
  Room:
    type: "object"
    required:
    - "reason"
    properties:
      id:
        type: "integer"
        format: "int64"
      roomId:
        type: "string"
        description: "Alfanumérico que identifica la sala (lo generá el backend)."
      reason:
        type: "string"
        description: "Motivo por el cual se crea la sala."
      createdDate:
        type: "string"
        format: "date-time"
        description: "Fecha y hora en la que se crea la sala."
      createdBy:
        type: "string"
        description: "Usuario que crea la sala."
      modifiedDate:
        type: "string"
        format: "date-time"
        description: "Fecha y hora en la que se modifica la sala."
      modifiedBy:
        type: "string"
        description: "Usuario que modifica la sala."
  UpdateRoom:
    type: "object"
    required:
    - "reason"
    properties:
      id:
        type: "integer"
        format: "int64"
      reason:
        type: "string"
        description: "Motivo por el cual se crea la sala."
      createdDate:
        type: "string"
        format: "date-time"
        description: "Fecha y hora en la que se crea la sala."
      createdBy:
        type: "string"
        description: "Usuario que crea la sala."
      modifiedDate:
        type: "string"
        format: "date-time"
        description: "Fecha y hora en la que se modifica la sala."
      modifiedBy:
        type: "string"
        description: "Usuario que modifica la sala."
  FormatError:
    type: "object"
    properties:
      timestamp:
        type: "string"
        format: "date-time"
      status:
        type: "integer"
        format: "int32"
      error:
        type: "string"
      trace:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"